Reorder home page sections and implement alternating backgrounds - resolves #401
- Moved "For Teachers" and "For Students" sections above Featured Modules as requested
- Extracted sections into standalone components for better maintainability:
  * ForTeachersSection.jsx - "For Teachers" section with gray background
  * ForStudentsSection.jsx - "For Students" section with white background
  * TestimonialsSection.jsx - "Testimonials" section with white background
- Updated HomePage.jsx to render sections in correct order:
  1. Hero Section
  2. For Teachers (gray background)
  3. For Students (white background)
  4. Featured Modules (gray background)
  5. Testimonials (white background)
- Implemented alternating background colors for visual contrast
- Removed "Learn More" and "Browse Content" buttons from section headers
- Standardized section padding to 60px for consistent spacing
- Cleaned up ExploreModulesSection.jsx to focus only on Featured Modules
- Maintained all existing functionality and conditional rendering logic
