+++ portal-app/src/components/ExploreModulesSection.jsx
@@ -2,13 +2,9 @@ import React, { useState, useEffect } from "react";
 import aiExploreImg from "../assets/ChatGPT Image Jun 13, 2025, 02_04_24 PM.png";
 import aiExploreImg2 from "../assets/ChatGPT Image Jun 13, 2025, 02_17_05 PM.png";
 import aiExploreImg3 from "../assets/ChatGPT Image Jun 13, 2025, 02_25_51 PM.png";
-import barchartImg from "../assets/barchart.png";
 import laptopImg from "../assets/laptop.png";
-import teacherImg from "../assets/teacher.png";
 import physicsImg from "../assets/finphysics.png";
 import textbooksImg from "../assets/textbooks.png";
-import microscopeImg from "../assets/microscope.png";
-import pencilImg from "../assets/finpencil.png";
 import { getFirestore, collection, getDocs, doc, onSnapshot } from "firebase/firestore";
 import { app as firebaseApp } from "../firebase/firebaseConfig";
 import { db } from "../firebase/firebaseConfig";
@@ -96,42 +92,6 @@ function capitalizeWords(str) {
 }
 
 
-// --- Sign Up Prompt Modal ---
-const SignUpPrompt = ({ open, onClose, type }) => {
-  if (!open) return null;
-  const isTeacher = type === "teacher";
-  return (
-    <div style={{
-      position: "fixed", top: 0, left: 0, right: 0, bottom: 0,
-      background: "rgba(0,0,0,0.3)", zIndex: 2000,
-      display: "flex", alignItems: "center", justifyContent: "center"
-    }}>
-      <div style={{
-        background: "#fff", borderRadius: 12, padding: 32, minWidth: 320,
-        boxShadow: "0 4px 24px rgba(0,0,0,0.18)", textAlign: "center", position: "relative"
-      }}>
-        <button onClick={onClose} style={{
-          position: "absolute", top: 10, right: 16, background: "none", border: "none",
-          fontSize: "1.5rem", cursor: "pointer", color: "#888"
-        }}>×</button>
-        <h3 style={{ marginBottom: 16 }}>
-          {isTeacher ? "Sign Up for Teacher Account" : "Sign Up for Student Account"}
-        </h3>
-        <div style={{ marginBottom: 24 }}>
-          Please sign up or log in to access this page.
-        </div>
-        <a href="/signup">
-          <button style={{
-            background: "#162040", color: "#fff", border: "none", borderRadius: 6,
-            padding: "12px 32px", fontWeight: 600, fontSize: "1rem", cursor: "pointer"
-          }}>
-            {isTeacher ? "Sign Up as Teacher" : "Sign Up as Student"}
-          </button>
-        </a>
-      </div>
-    </div>
-  );
-};
 
 // --- Custom Hook to get user and role from Firebase ---
 function useUserRole() {
@@ -175,748 +135,10 @@ function useUserRole() {
   return { user, role };
 }
 
-// --- SquareSection Component ---
-const SquareSection = ({ title, description, buttonText, buttonLink, children }) => (
-  <section
-    style={{
-      width: "100%",
-      background: "#F6F8FA",
-      padding: "60px 0 0 0",
-      display: "flex",
-      flexDirection: "column",
-      alignItems: "center",
-      justifyContent: "flex-start"
-    }}
-  >
-    <h2
-      style={{
-        fontSize: "2.5rem",
-        fontWeight: "700",
-        color: "#111",
-        fontFamily: "Open Sans, sans-serif",
-        textAlign: "center",
-        margin: 0,
-        letterSpacing: "1px"
-      }}
-    >
-      {title}
-    </h2>
-    {description && (
-      <p
-        style={{
-          marginTop: "18px",
-          fontSize: "1.15rem",
-          color: "#222",
-          textAlign: "center",
-          maxWidth: "600px",
-          fontWeight: 500,
-        }}
-      >
-        {description}
-      </p>
-    )}
-    {buttonText && (
-      <button
-        style={{
-          marginTop: "32px",
-          background: "#162040",
-          color: "#fff",
-          border: "2px solid #162040",
-          borderRadius: "6px",
-          padding: "14px 48px",
-          fontSize: "1.08rem",
-          fontWeight: "600",
-          cursor: "pointer",
-          transition: "background 0.2s, color 0.2s, border 0.2s",
-          minWidth: "260px",
-        }}
-        onClick={() => window.location.href = buttonLink || "#"}
-      >
-        {buttonText}
-      </button>
-    )}
-    {children}
-  </section>
-);
-
-// --- TeacherRectangles with role check ---
-const TeacherRectangles = () => {
-  const { user, role } = useUserRole();
-  const [showPrompt, setShowPrompt] = useState(false);
-
-  const handleClick = (e) => {
-    if (
-      !user ||
-      !["teacherDefault", "teacherPlus", "admin"].includes(role)
-    ) {
-      e.preventDefault();
-      setShowPrompt(true);
-    }
-  };
-
-  return (
-    <div
-      style={{
-        display: "flex",
-        flexWrap: "wrap",
-        justifyContent: "center",
-        gap: "40px",
-        marginTop: "48px",
-        marginBottom: "32px",
-        width: "100%",
-        maxWidth: "1100px"
-      }}
-    >
-      {/* Rectangle 1 */}
-      <a href="/modules" style={{ textDecoration: "none" }} onClick={handleClick}>
-        <div
-          style={{
-            background: "#f3f3f1",
-            borderRadius: "12px",
-            boxShadow: "0 2px 8px rgba(0,0,0,0.06)",
-            border: "1px solid #e0dfdb",
-            width: "480px",
-            minHeight: "160px",
-            padding: "0",
-            display: "flex",
-            flexDirection: "row",
-            alignItems: "center",
-            cursor: "pointer"
-          }}
-        >
-          <div style={{
-            width: "90px",
-            height: "90px",
-            marginLeft: "32px",
-            marginRight: "24px",
-            borderRadius: "8px",
-            background: "#e0dfdb",
-            display: "flex",
-            alignItems: "center",
-            justifyContent: "center",
-            overflow: "hidden"
-          }}>
-            <img
-              src={barchartImg}
-              alt="Barchart"
-              style={{
-                width: "80px",
-                height: "80px",
-                objectFit: "contain",
-                display: "block"
-              }}
-            />
-          </div>
-          <div>
-            <span style={{ fontWeight: "700", fontSize: "1.25rem", color: "#162040" }}>
-              Ready-to-use Modules
-            </span>
-            <div style={{ marginTop: "10px", color: "#222", fontSize: "1.08rem", maxWidth: "280px" }}>
-              Access a library of pre-built modules.
-            </div>
-          </div>
-        </div>
-      </a>
-      {/* Rectangle 2 */}
-      <a href="/lesson-plans" style={{ textDecoration: "none" }} onClick={handleClick}>
-        <div
-          style={{
-            background: "#f3f3f1",
-            borderRadius: "12px",
-            boxShadow: "0 2px 8px rgba(0,0,0,0.06)",
-            border: "1px solid #e0dfdb",
-            width: "480px",
-            minHeight: "160px",
-            padding: "0",
-            display: "flex",
-            flexDirection: "row",
-            alignItems: "center",
-            cursor: "pointer"
-          }}
-        >
-          <div style={{
-            width: "90px",
-            height: "90px",
-            marginLeft: "32px",
-            marginRight: "24px",
-            borderRadius: "8px",
-            background: "#e0dfdb",
-            display: "flex",
-            alignItems: "center",
-            justifyContent: "center",
-            overflow: "hidden"
-          }}>
-            <img
-              src={laptopImg}
-              alt="Laptop"
-              style={{
-                width: "100%",
-                height: "100%",
-                objectFit: "cover",
-                display: "block"
-              }}
-            />
-          </div>
-          <div>
-            <span style={{ fontWeight: "700", fontSize: "1.25rem", color: "#162040" }}>
-              Lesson Plan Builder
-            </span>
-            <div style={{ marginTop: "10px", color: "#222", fontSize: "1.08rem", maxWidth: "280px" }}>
-              Create and customize your lesson plans.
-            </div>
-          </div>
-        </div>
-      </a>
-      {/* Rectangle 3 */}
-      <a href="/classroom-management" style={{ textDecoration: "none" }} onClick={handleClick}>
-        <div
-          style={{
-            background: "#f3f3f1",
-            borderRadius: "12px",
-            boxShadow: "0 2px 8px rgba(0,0,0,0.06)",
-            border: "1px solid #e0dfdb",
-            width: "480px",
-            minHeight: "160px",
-            padding: "0",
-            display: "flex",
-            flexDirection: "row",
-            alignItems: "center",
-            cursor: "pointer"
-          }}
-        >
-          <div style={{
-            width: "90px",
-            height: "90px",
-            marginLeft: "32px",
-            marginRight: "24px",
-            borderRadius: "8px",
-            background: "#e0dfdb",
-            display: "flex",
-            alignItems: "center",
-            justifyContent: "center",
-            overflow: "hidden"
-          }}>
-            <img
-              src={teacherImg}
-              alt="Teacher"
-              style={{
-                width: "80px",
-                height: "80px",
-                objectFit: "contain",
-                display: "block"
-              }}
-            />
-          </div>
-          <div>
-            <span style={{ fontWeight: "700", fontSize: "1.25rem", color: "#162040" }}>
-              Classroom Management
-            </span>
-            <div style={{ marginTop: "10px", color: "#222", fontSize: "1.08rem", maxWidth: "280px" }}>
-              Control content visibility for students.
-            </div>
-          </div>
-        </div>
-      </a>
-      {/* Rectangle 4 */}
-      <a href="/community" style={{ textDecoration: "none" }} onClick={handleClick}>
-        <div
-          style={{
-            background: "#f3f3f1",
-            borderRadius: "12px",
-            boxShadow: "0 2px 8px rgba(0,0,0,0.06)",
-            border: "1px solid #e0dfdb",
-            width: "480px",
-            minHeight: "160px",
-            padding: "0",
-            display: "flex",
-            flexDirection: "row",
-            alignItems: "center",
-            cursor: "pointer"
-          }}
-        >
-          <div style={{
-            width: "90px",
-            height: "90px",
-            marginLeft: "32px",
-            marginRight: "24px",
-            borderRadius: "8px",
-            background: "#e0dfdb",
-            display: "flex",
-            alignItems: "center",
-            justifyContent: "center",
-            overflow: "hidden"
-          }}>
-            <img
-              src={pencilImg}
-              alt="Pencil"
-              style={{
-                width: "80px",
-                height: "80px",
-                objectFit: "contain",
-                display: "block"
-              }}
-            />
-          </div>
-          <div>
-            <span style={{ fontWeight: "700", fontSize: "1.25rem", color: "#162040" }}>
-              Share with Community
-            </span>
-            <div style={{ marginTop: "10px", color: "#222", fontSize: "1.08rem", maxWidth: "280px" }}>
-              Make the lesson plan public to the community.
-            </div>
-          </div>
-        </div>
-      </a>
-      <SignUpPrompt open={showPrompt} onClose={() => setShowPrompt(false)} type="teacher" />
-    </div>
-  );
-};
-
-// --- StudentRectangles with role check ---
-const StudentRectangles = () => {
-  const { user, role } = useUserRole();
-  const [showPrompt, setShowPrompt] = useState(false);
-
-  const handleClick = (e) => {
-    if (
-      !user ||
-      !["student", "consumer"].includes(role)
-    ) {
-      e.preventDefault();
-      setShowPrompt(true);
-    }
-  };
-
-  return (
-    <div
-      style={{
-        display: "flex",
-        flexWrap: "nowrap",
-        justifyContent: "center",
-        gap: "40px",
-        marginTop: "48px",
-        marginBottom: "32px",
-        width: "100%",
-        maxWidth: "1100px"
-      }}
-    >
-      {/* Rectangle 1 */}
-      <a href="/learning-modules" style={{ textDecoration: "none" }} onClick={handleClick}>
-        <div
-          style={{
-            background: "#f3f3f1",
-            borderRadius: "12px",
-            boxShadow: "0 2px 8px rgba(0,0,0,0.06)",
-            border: "1px solid #e0dfdb",
-            width: "480px",
-            minHeight: "160px",
-            padding: "0",
-            display: "flex",
-            flexDirection: "row",
-            alignItems: "center",
-            cursor: "pointer"
-          }}
-        >
-          <div style={{
-            width: "90px",
-            height: "90px",
-            marginLeft: "32px",
-            marginRight: "24px",
-            borderRadius: "8px",
-            background: "#e0dfdb",
-            display: "flex",
-            alignItems: "center",
-            justifyContent: "center",
-            overflow: "hidden"
-          }}>
-            <img
-              src={textbooksImg}
-              alt="Textbooks"
-              style={{
-                width: "80px",
-                height: "80px",
-                objectFit: "contain",
-                display: "block"
-              }}
-            />
-          </div>
-          <div>
-            <span style={{ fontWeight: "700", fontSize: "1.25rem", color: "#162040" }}>
-              Learning Modules
-            </span>
-            <div style={{ marginTop: "10px", color: "#222", fontSize: "1.08rem", maxWidth: "280px" }}>
-              Interactive content to enhance your understanding.
-            </div>
-          </div>
-        </div>
-      </a>
-      {/* Rectangle 2 */}
-      <a href="/project-ideas" style={{ textDecoration: "none" }} onClick={handleClick}>
-        <div
-          style={{
-            background: "#f3f3f1",
-            borderRadius: "12px",
-            boxShadow: "0 2px 8px rgba(0,0,0,0.06)",
-            border: "1px solid #e0dfdb",
-            width: "480px",
-            minHeight: "160px",
-            padding: "0",
-            display: "flex",
-            flexDirection: "row",
-            alignItems: "center",
-            cursor: "pointer"
-          }}
-        >
-          <div style={{
-            width: "90px",
-            height: "90px",
-            marginLeft: "32px",
-            marginRight: "24px",
-            borderRadius: "8px",
-            background: "#e0dfdb",
-            display: "flex",
-            alignItems: "center",
-            justifyContent: "center",
-            overflow: "hidden"
-          }}>
-            <img
-              src={microscopeImg}
-              alt="Microscope"
-              style={{
-                width: "80px",
-                height: "80px",
-                objectFit: "contain",
-                display: "block"
-              }}
-            />
-          </div>
-          <div>
-            <span style={{ fontWeight: "700", fontSize: "1.25rem", color: "#162040" }}>
-              Project Ideas for Science Fair
-            </span>
-            <div style={{ marginTop: "10px", color: "#222", fontSize: "1.08rem", maxWidth: "280px" }}>
-              Get inspired with creative project ideas.
-            </div>
-          </div>
-        </div>
-      </a>
-      <SignUpPrompt open={showPrompt} onClose={() => setShowPrompt(false)} type="student" />
-    </div>
-  );
-};
 
 
-const PopupTestimonialCard = ({
-  testimonial,
-  open,
-  onClick,
-  onClose,
-  forcePreview,
-  onPopupPrev,
-  onPopupNext,
-  showPopupNav,
-  isFirst,
-  isLast
-}) => {
-  const previewText = forcePreview || testimonial.Text;
 
-  // Only render the preview card if onClick is provided (carousel preview)
-  if (!open && onClick) {
-    return (
-      <div
-        onClick={onClick}
-        style={{
-          background: "#f3f3f1",
-          borderRadius: "10px",
-          padding: "24px 28px",
-          maxWidth: "260px",
-          minWidth: "220px",
-          minHeight: "250px",
-          height: "300px",
-          boxSizing: "border-box",
-          boxShadow: "0 1px 4px rgba(0,0,0,0.04)",
-          border: "1px solid #e0dfdb",
-          display: "flex",
-          flexDirection: "column",
-          justifyContent: "space-between",
-          cursor: "pointer",
-          position: "relative",
-          zIndex: 1,
-        }}
-      >
-        <div
-          style={{
-            color: "#222",
-            fontSize: "1.05rem",
-            marginBottom: "12px",
-            flex: 1,
-            overflow: "hidden",
-            lineHeight: 1.5,
-            maxHeight: "200px",
-            minHeight: "60px",
-            transition: "opacity 0.2s"
-          }}
-        >
-          {`“${previewText}”`}
-        </div>
-        <div style={{
-          color: "#111",
-          fontSize: "0.98rem",
-          fontWeight: "600",
-          marginTop: "8px"
-        }}>
-          – {testimonial.Name}
-        </div>
-      </div>
-    );
-  }
 
-  // Only render the popup if open
-  if (!open) return null;
-
-  return (
-    <div
-      onClick={onClose}
-      style={{
-        position: "fixed",
-        top: 0, left: 0, right: 0, bottom: 0,
-        background: "rgba(0,0,0,0.3)",
-        zIndex: 1000,
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      }}
-    >
-      <div
-        onClick={e => e.stopPropagation()}
-        style={{
-          background: "#fff",
-          borderRadius: "14px",
-          padding: "32px 36px 24px 36px",
-          maxWidth: "400px",
-          width: "90%",
-          boxShadow: "0 4px 24px rgba(0,0,0,0.18)",
-          position: "relative",
-          display: "flex",
-          flexDirection: "column",
-          alignItems: "center",
-          maxHeight: "80vh",         // <-- Add this
-          overflowY: "auto"          // <-- And this
-        }}
-      >
-        {/* Popup navigation arrows at the top, but lower so they don't cover the exit button */}
-        {showPopupNav && (
-          <div style={{
-            display: "flex",
-            justifyContent: "space-between",
-            alignItems: "center",
-            position: "absolute",
-            top: 40, // Slightly higher than before (was 54)
-            left: 0,
-            right: 0,
-            width: "100%",
-            padding: "0 16px",
-            zIndex: 2
-          }}>
-            <button
-              onClick={onPopupPrev}
-              disabled={isFirst}
-              style={{
-                background: "#fff",
-                border: "1px solid #ccc",
-                borderRadius: "50%",
-                width: 32,
-                height: 32,
-                fontSize: "1.3rem",
-                color: "#162040",
-                cursor: isFirst ? "not-allowed" : "pointer",
-                opacity: isFirst ? 0.4 : 1,
-                transition: "opacity 0.2s"
-              }}
-              aria-label="Previous testimonial"
-            >
-              &#8592;
-            </button>
-            <button
-              onClick={onPopupNext}
-              disabled={isLast}
-              style={{
-                background: "#fff",
-                border: "1px solid #ccc",
-                borderRadius: "50%",
-                width: 32,
-                height: 32,
-                fontSize: "1.3rem",
-                color: "#162040",
-                cursor: isLast ? "not-allowed" : "pointer",
-                opacity: isLast ? 0.4 : 1,
-                transition: "opacity 0.2s"
-              }}
-              aria-label="Next testimonial"
-            >
-              &#8594;
-            </button>
-          </div>
-        )}
-        <button
-          onClick={onClose}
-          style={{
-            position: "absolute",
-            top: 12,
-            right: 18,
-            background: "none",
-            border: "none",
-            fontSize: "1.5rem",
-            cursor: "pointer",
-            color: "#888",
-            zIndex: 3
-          }}
-          aria-label="Close"
-        >
-          ×
-        </button>
-        <div style={{ color: "#222", fontSize: "1.08rem", marginBottom: "18px", marginTop: showPopupNav ? "38px" : "0", textAlign: "center" }}>
-          {`“${testimonial.Text}”`}
-        </div>
-        <div style={{ color: "#111", fontSize: "1rem", fontWeight: 600, textAlign: "center" }}>
-          – {testimonial.Name}
-        </div>
-      </div>
-    </div>
-  );
-};
-
-// In TestimonialsCarousel, pass popup navigation handlers and state
-const TestimonialsCarousel = () => {
-  const [testimonials, setTestimonials] = useState([]);
-  const [openIndex, setOpenIndex] = useState(null);
-  const [startIndex, setStartIndex] = useState(0);
-
-  // Fetch testimonials from Firestore on mount
-  useEffect(() => {
-    const fetchTestimonials = async () => {
-      const db = getFirestore(firebaseApp);
-      const querySnapshot = await getDocs(collection(db, "testimonials"));
-      const data = [];
-      querySnapshot.forEach((doc) => {
-        data.push({ id: doc.id, ...doc.data() });
-      });
-      // Sort: Teachers first, then others
-      data.sort((a, b) => {
-        const roleA = (a.Role || a.role || "").toLowerCase();
-        const roleB = (b.Role || b.role || "").toLowerCase();
-        if (roleA === "teacher" && roleB !== "teacher") return -1;
-        if (roleA !== "teacher" && roleB === "teacher") return 1;
-        return 0;
-      });
-      setTestimonials(data);
-    };
-    fetchTestimonials();
-  }, []);
-
-  // Show 2 testimonials at a time
-  const testimonialsPerPage = 2;
-  const maxIndex = Math.max(0, testimonials.length - testimonialsPerPage);
-
-  const handlePrev = () => setStartIndex(i => (i > 0 ? Math.max(0, i - testimonialsPerPage) : i));
-  const handleNext = () => setStartIndex(i => (i < maxIndex ? Math.min(maxIndex, i + testimonialsPerPage) : i));
-
-  const visibleTestimonials = testimonials.slice(startIndex, startIndex + testimonialsPerPage);
-
-  // Optionally, you can generate previews if you want to keep the preview logic
-  const previews = visibleTestimonials.map(t =>
-    t.Text.length > 120 ? t.Text.slice(0, 120) + "..." : t.Text
-  );
-
-  // Popup navigation handlers
-  const handlePopupPrev = () => {
-    if (openIndex > 0) setOpenIndex(openIndex - 1);
-  };
-  const handlePopupNext = () => {
-    if (openIndex < testimonials.length - 1) setOpenIndex(openIndex + 1);
-  };
-
-  return (
-    <div style={{
-      flex: 1,
-      display: "flex",
-      alignItems: "center",
-      justifyContent: "flex-end",
-      minWidth: 0,
-      marginLeft: "40px",
-      gap: "24px"
-    }}>
-      <button
-        onClick={handlePrev}
-        disabled={startIndex === 0}
-        style={{
-          background: "#fff",
-          border: "1px solid #ccc",
-          borderRadius: "50%",
-          width: 36,
-          height: 36,
-          marginRight: 12,
-          fontSize: "1.5rem",
-          color: "#162040",
-          cursor: startIndex === 0 ? "not-allowed" : "pointer",
-          opacity: startIndex === 0 ? 0.4 : 1,
-          transition: "opacity 0.2s"
-        }}
-        aria-label="Previous testimonials"
-      >
-        &#8592;
-      </button>
-      {/* Only render preview cards for visibleTestimonials */}
-      {visibleTestimonials.map((testimonial, idx) => {
-        const globalIdx = startIndex + idx;
-        return (
-          <PopupTestimonialCard
-            key={testimonial.id || globalIdx}
-            testimonial={testimonial}
-            open={false}
-            onClick={() => setOpenIndex(globalIdx)}
-            forcePreview={previews[idx]}
-          />
-        );
-      })}
-      <button
-        onClick={handleNext}
-        disabled={startIndex >= maxIndex}
-        style={{
-          background: "#fff",
-          border: "1px solid #ccc",
-          borderRadius: "50%",
-          width: 36,
-          height: 36,
-          marginLeft: 12,
-          fontSize: "1.5rem",
-          color: "#162040",
-          cursor: startIndex >= maxIndex ? "not-allowed" : "pointer",
-          opacity: startIndex >= maxIndex ? 0.4 : 1,
-          transition: "opacity 0.2s"
-        }}
-        aria-label="Next testimonials"
-      >
-        &#8594;
-      </button>
-
-      {/* Popups for all testimonials, only one is open at a time */}
-      {testimonials.map((testimonial, idx) => (
-        <PopupTestimonialCard
-          key={testimonial.id || idx}
-          testimonial={testimonial}
-          open={openIndex === idx}
-          onClose={() => setOpenIndex(null)}
-          onPopupPrev={handlePopupPrev}
-          onPopupNext={handlePopupNext}
-          showPopupNav={openIndex !== null}
-          isFirst={openIndex === 0}
-          isLast={openIndex === testimonials.length - 1}
-        />
-      ))}
-    </div>
-  );
-};
 
 function ModuleLoginPrompt({ open, onClose, moduleTitle, summary }) {
   if (!open) return null;
@@ -1388,7 +610,7 @@ const ExploreModulesSection = () => {
       <section
         style={{
           width: "100%",
-          padding: "100px 0 0 0",
+          padding: "60px 0 60px 0",
           minHeight: "700px",
           display: "flex",
           flexDirection: "column",
@@ -2038,97 +1260,6 @@ const ExploreModulesSection = () => {
         )}
       </section>
 
-      {/* Only show For Teachers if NOT a student and NOT teacherDefault */}
-      {!isTeacherDefault && (!role || !["student", "consumer"].includes(role)) && (
-        <div style={{ width: "100%", marginTop: "-40px" }}>
-          <SquareSection
-            title="For Teachers"
-            description="Unlock powerful tools to enhance your teaching."
-            buttonText="Learn More"
-            buttonLink="/teachers"
-          >
-            <TeacherRectangles />
-          </SquareSection>
-        </div>
-      )}
-
-      {/* Only show For Students if NOT a teacher */}
-      {(!role || !["teacherDefault", "teacherPlus", "admin"].includes(role)) && (
-        <div style={{ width: "100%" }}>
-          <SquareSection
-            title="For Students"
-            description="Discover engaging content tailored for your learning."
-            buttonText="Browse Content"
-            buttonLink="/students"
-          >
-            <StudentRectangles />
-          </SquareSection>
-        </div>
-      )}
-
-      {/* Add spacing between For Students and Reviews section */}
-      <div style={{ height: "60px" }} />
-
-      {/* Only show Testimonials if NOT teacherDefault */}
-      {!isTeacherDefault && (
-        <section
-          style={{
-            width: "100%",
-            background: "#F6F8FA",
-            padding: "0",
-            display: "flex",
-            flexDirection: "row",
-            alignItems: "flex-start",
-            justifyContent: "center",
-            gap: "80px",
-            maxWidth: "1100px",
-            margin: "0 auto 60px auto"
-          }}
-        >
-          {/* Left: Title and Description */}
-          <div style={{
-            flex: "0 0 260px",
-            display: "flex",
-            flexDirection: "column",
-            alignItems: "flex-start",
-            justifyContent: "center",
-            gap: "16px",
-            maxWidth: "100%"
-          }}>
-            <h2 style={{
-              fontSize: "2.5rem",
-              fontWeight: "700",
-              color: "#162040",
-              marginBottom: "18px",
-              textAlign: "center",
-              fontFamily: "Open Sans, sans-serif",
-              letterSpacing: "1px",
-              whiteSpace: "nowrap",
-              overflow: "hidden",
-              textOverflow: "ellipsis"
-            }}>
-              Testimonials
-            </h2>
-            <p
-              style={{
-                fontSize: "1.15rem",
-                color: "#222",
-                maxWidth: "1000px",
-                lineHeight: 1.6,
-                margin: 0,
-                textAlign: "center"
-              }}
-            >
-              Discover how our platform has transformed the teaching and learning experience for educators and students alike.
-            </p>
-          </div>
-
-          {/* Right: Testimonials Carousel */}
-          <div style={{ flex: "1 1 300px", position: "relative", width: "100%" }}>
-            <TestimonialsCarousel />
-          </div>
-        </section>
-      )}
       <UpgradePrompt open={upgradePromptOpen} onClose={() => setUpgradePromptOpen(false)} />
     </div>
   );
