+++ portal-app/src/components/HomePage.jsx
@@ -1,21 +1,21 @@
-import React, { useState } from "react";
-import content_creator from "../assets/content_creator.png";
-import educator from "../assets/educator.png";
-import lesson_plans from "../assets/lesson_plans.png";
-import { useNavigate } from "react-router-dom";
+import React from "react";
 import useUserData from "../hooks/useUserData";
 import HeroSection from "../components/HeroSection";
+import ForTeachersSection from "../components/ForTeachersSection";
+import ForStudentsSection from "../components/ForStudentsSection";
 import ExploreModulesSection from "../components/ExploreModulesSection";
+import TestimonialsSection from "../components/TestimonialsSection";
 
 const HomePage = ({ user }) => {
-  const navigate = useNavigate();
   const { userData } = useUserData();
-  const [showUpgradePopup, setShowUpgradePopup] = useState(false);
 
   return (
     <>
       <HeroSection />
+      <ForTeachersSection />
+      <ForStudentsSection />
       <ExploreModulesSection userData={userData} />
+      <TestimonialsSection />
       {/* Footer or any content you want to keep below */}
     </>
   );
